pyproject.toml
[tool.poetry]
name = "abner"
version = "0.1.0"
description = ""
authors = ["Connor McLean <c.mclean165@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
lasio = "^0.31"
scipy = "^1.14.1"
wikipedia = "^1.4.0"
speechrecognition = "^3.11.0"
pyttsx3 = "^2.98"
google-cloud-speech = "^2.28.1"
google-api-python-client = "^2.154.0"
tk = "^0.1.0"
pillow = "^11.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
